pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "silver"
version = "0.1.0"
description = "Point-in-time prediction and backtesting system for US equities."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["PyYAML>=6.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"
src = ["src", "tests"]
